flask-backend/requirements.txt
Flask
Flask-SQLAlchemy
flask-cors
psycopg2-binary
google-auth
google-auth-oauthlib
google-auth-httplib2
